Split tasks of PhysicalFixedBatchCopy into a separate class (BatchTaskHelper)
